Adjusted DustFX to work properly on Lander
Changed the DustFX sub object into its own prefab and used nested prefabs with it in the mothership and on the lander. Hooked up special functionality to handle the dust with the thrust controller on the Lander (moves the fx spawner to the correct location when needed and uses a 2D capsule collider instead of sphere as it is closer shape to the thrust graphic. Moved Lander, mothership and main camera into a Player folder under Prefabs (to tidy up)